Garbage collection!
A wasteful breath-first copying algorithm. Inspired by Cheney's
algorithm as described in section 16.2 of Appel's Compiling with
Continuations. But less elegant and simple, as I need a separate queue
for the scan in order to know the type of the objects in to-space
still to be processed.

Need to investigate why some tests can still not be bumped up to their
pre-CPS-conversion efficiency.