Step 7: Heap allocation
This is a very rudimentary start. Allocating the heap in the main's stack lead
to huge gaps in the stack which I don't understand yet. At least this is simpler
to reason about (small stack) and debug (very different address, no inline asm
to initialize an array).

As of now, the starting address of the heap is available at `$rsp+32` after the
function preamble. I can't say that I understand all of this yet.