Fix argument calling conventions
This might be a small patch, but a lot went into it. It took a few days of
reading and research to figure out the default calling convention and how
arguments are passed into it. After fiddling with the stack a lot, it turned out
to be that all arguments were passed in the registers as per the System V
convention.

- We allocate a register RSI to keep track of the top of the heap.
- Fix the extern declaration
- Use 64bit arguments everywhere