Pairs in heap! Wo0t!
Alright, too much excitement. Pairs are allocated in heap and slightly non
trivial examples seems to be working fine. A pair takes 2 words in the heap and
the result is the base address is tagged with `pair-tag`.

Consider following example `(cons 3 (cons 2 1))` with a base heap address 800.
Arguments to cons are evaluated and stored temporarily in stack. `(cons 2 1)`
modifies the heap to

    | 2          | 1         |
    ^   ^        ^           ^
    800 801      808         816 <- RSI

The base address of the heap is tagged (800 | 1) and the result will be 801. The
outer expression modifies the heap again to

    | 2          | 1         | 3         | 801        |
    ^   ^        ^           ^                        ^
    800 801      808         816                      832 <- RSI

`car` of a pair can be accessed at `address - 1` and `cdr` at `address + 7`.