Simple C style top level functions with letrec
This patch probably needed more work than the rest of the project combined.

The top level letrec adds a jump target for each function with the bound name.
Memory addresses are aligned with RBP instead of RSP to make it easier to use
push and pop instructions. Several things like stack handling inside boolean
expressions is still messed up, but this calls for :v: