Pairs in heap again! Wo0t!
A 'heap' is just a continuous range of memory allocated once at program start
and a reference is maintained at`RSI` at all the times. A pair allocates both
arguments in this memory block adjacent to each other, increments `RSI` by 2
words and tags the base address with `PAIRTAG`.

Slightly non trivial examples seems to be working fine; should probably add a
couple of quick checks as well.

Consider example `(cons 3 (cons 2 1))` with a base heap address 800. Arguments
to cons are evaluated and pushed to heap. `(cons 2 1)` modifies the heap to

        | 2          | 1         |
        ^   ^        ^           ^
        800 801      808         816 <- RSI

The base address of the heap is tagged (800 | 3) and represents the first pair.
The outer expression modifies the heap again to

        | 2          | 1         | 3         | 803        |
        ^   ^        ^           ^                        ^
        800 801      808         816                      832 <- RSI

`car` of a pair can be accessed at `address - 3` and `cdr` at `address + 5`.

See 9786392c for the first scheme implementation