Add rudimentary static string support 🏆🥇
The original scheme version was pretty crappy because it immediate encoded every
character. This would make almost all operations on strings harder and extremely
inefficient, would make unicode support impossible, makes generating the asm
quite painful and every FFI extremely painful - there was pretty much nothing
good about it. I'm reasonably happy about this v2!

Strings are a pair of length and a pointer to a blob of UTF-8 encoded bytes.

The raw data bytes can be allocated in stack, heap or statically in the
generated binary. Static strings found in the source code is retained as it
is in the generated binary and a label is used for dereference. Strings
generated at runtime must be allocated in the heap.

Example memory layout:

A literal "hello world" gets statically allocated at offset 4000. A string
object will be allocated in the heap at address 8000 as a pair `(11, 4000)`.
A reference to this object as a local variable in stack at address 12000
would be an immediately tagged pointer 8005. `(8000 | string tag)`

```txt
 -------------------------
| Address | Value         |
 -------------------------
| 4000    | "hello world" |
|         |               |
| 8000    | 11            |
| 8008    | 4000          |
|         |               |
| 12000   | 8005          |
 -------------------------
```

The C runtime would get the value 8005 and would identify it as a string
`(8005 & mask == strtag)`. The raw pointer is obtained by removing the tag
`(p = val - strtag)` and length is found at the base addresses `(*p)` and
the pointer to raw blob is found at the next addresses `(*p + 1)`. fwrite
can safely print the exact number of bytes using the length and pointer.