Remove an unnecessary pointer indirection in strings
A string object returned to runtime used to be a pointer to a pointer but now
its only a single pointer! Woo

The essence is lost in the patch noise, but the runtime diff really shines!

```
-        int64_t str = *(p + 1);
+        int64_t *str = (p + 1);
```

The prior version allocated the raw bytes in data section, allocated a pair with
the length and pointer to the data in heap and then immediate tagged this
object. This patch removes one level of indirection by removing the heap
allocated pair. The length of the blob is prefixed before the data and the
static address is directly tagged.

Current and previous code side by side for comparison:

```
Current version                       Previous version

_init:                                _init:
    ...                                   ...
    lea rax, [rip + 5 + _inc_str_0]       lea rdi, [rip + _inc_str_0]
    pop rbp                               mov qword ptr [rsi + 0], 11
    ret                                   mov qword ptr [rsi + 8], rdi
                                          mov rax, rsi
    .p2align 3                            or rax, 5
_inc_str_0:                               add rsi, 16
    .quad  11                             pop rbp
    .ascii "hello world"                  ret

                                      _inc_str_0:
                                          .ascii "hello world"
```