Add support for the simplest C style functions 🦑
Its a mess due to the odd calling convention, but that can be cleaned up later!
Switching to System V AMD64 which passes all args in registers would make this
much better.

The patch that did the same in Scheme ddcd125270d171545085052d3629ab25fae7e55a
took more effort than the rest of the project combined and this one turned out
to be the same again.

Sample:

```
$ echo -n "(let ((x 42) (id (lambda (x) x))) (id x))" | cargo run -q -- -S

    .section __TEXT,__text
    .intel_syntax noprefix

    .globl _init
_init:
    push rbp
    mov rbp, rsp
    mov rsi, rdi        # Store heap index to RSI
    mov rax, 336
    mov [rbp - 8], rax
    mov rax, [rbp - 8]
    mov [rbp - 32], rax
    add rsp, -8
    call id
    add rsp, 8
    pop rbp
    ret

    .globl id
id:
    push rbp
    mov rbp, rsp
    mov rax, [rbp - 8]
    pop rbp
    ret
```