✈ Replace `Addressable` trait object with an explicit `Reference` enum
Even though trait objects made the API a lot cleaner for the
callers, (x86::mov(RAX, 42) vs x86::mov(RAX.into(), 42.into()), it turned out to
be a complex thing under the hood to work with. Lifetimes for dynamic trait
objects are quite painful to work with and I couldn't make it work at all with
environments.

This is an explicit trade off of simplicity vs verbosity that can be revisited
later.

This patch was written on ✈ BA 49 from LHR to SEA in a pretty awful economy seat
packed like sardines. I think I've enough money to fly at least premium economy
from now on and I should probably do that. You remember how awful it is the
moment you get inside this silly tube.

So that's 2 lessons learned on a single flight.